base_requirements.txt
jangorestframework